fix(earthly): don't use vendored dependencies
Cargo exhibits subtle differences in behavior when dependencies are
vendored locally, which can sometimes lead to breakage in CI, where
the build suceeds on a developers local machine.

It is by no means obvious that the vendored dependencies are the cause
of such breakages when they occur, further adding to the headache.

For simplicity, and to more closely match a developers workflow,
we now simply fetch the dependencies, instead of vendoring. This allows
us to maintain the benefits of seperating the fetch phase from the
build (earthly can skip it if it hasn't changed), without the unexpected
breakage.