ci(refactor): generate weights in earthly
It is much more natural and less error prone to generate the weights
for production builds directly in Earthly itself. It also translates
to much less code.

This commit also simplifies GHCR image uploads to the registry, as
Earthly already has logic to tag an image with more than one tag, we
can just perform the upload directly with it.