chore: update polkadot-sdk dependency to polkadot-stable-2412 (#450)
Benchmarking command was removed because it is now obsolete way and frame-benchmarking-cli is GPLv3.

OffchainWorkers are not used by this node and polkadot-sdk code does not
give an option to not try load native certificates. Dependency that is
currently used for reading native certs fails on some light docker
images. So OffchainWorkers wiring in was removed.